Can now smelt Patchwork Flesh unconditionally
- Technically, from a user's perspective, you've always been able to
smelt Patchwork Flesh unconditionally, because there's no config option
to turn it off.
- But if the code that makes the 5x Rotten Flesh --> 1x Patchwork Flesh
recipes configurable were uncommented, then disabling it would have also
turned off the smelting recipe to smelt Patchwork Flesh into Leather,
leaving a useless item in the game if someone turned it off and then
loaded a world that used to have it on.  Useless items are bad.
- Since Patchwork Flesh is basically an IOU for 1 Leather, redeemable
for the cost of 1 Furnace action, it's not imbalanced to unconditionally
enable the smelting action if someone decides to disable the 5x Rotten
Flesh --> 1x Patchwork Flesh recipe.

Conflicts:
	source/airbreather/mods/yalsm/YalsmRecipeConfiguration.java